refactor: separate components into files
