+++ ReconnectED-Android/app/src/main/java/com/getreconnected/reconnected/MainActivity.kt
@@ -1,5 +1,6 @@
 package com.getreconnected.reconnected
 
+import android.os.Build
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
@@ -7,13 +8,17 @@ import androidx.activity.enableEdgeToEdge
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
+import com.getreconnected.reconnected.ui.composables.Greeting
 import com.getreconnected.reconnected.ui.theme.ReconnectEdTheme
 
+/**
+ * The main activity for the app.
+ */
 class MainActivity : ComponentActivity() {
+    /**
+     * Called when the activity is starting.
+     */
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
@@ -21,7 +26,7 @@ class MainActivity : ComponentActivity() {
             ReconnectEdTheme {
                 Scaffold(modifier = Modifier.fillMaxSize()) { innerPadding ->
                     Greeting(
-                        name = "Android",
+                        name = Build.USER,
                         modifier = Modifier.padding(innerPadding)
                     )
                 }
@@ -30,18 +35,3 @@ class MainActivity : ComponentActivity() {
     }
 }
 
-@Composable
-fun Greeting(name: String, modifier: Modifier = Modifier) {
-    Text(
-        text = "Hello $name!",
-        modifier = modifier
-    )
-}
-
-@Preview(showBackground = true)
-@Composable
-fun GreetingPreview() {
-    ReconnectEdTheme {
-        Greeting("Android")
-    }
-}
