+++ ReconnectED-Android/app/src/main/java/com/getreconnected/reconnected/core/Info.kt
@@ -1,0 +1,4 @@
+package com.getreconnected.reconnected.core
+
+val APPLICATION_NAME = "ReconnectED"
+val APPLICATION_DESCRIPTION = "Minimize the overconsumption of internet."
